requirements.txt
streamlit==1.15.0
openai==0.26.5

#asr whisper model
numpy
torch
tqdm
more-itertools
transformers>=4.19.0
ffmpeg-python==0.2.0
git+https://github.com/openai/whisper.git
soundfile==0.11.0

#audio recording related
python-magic==0.4.27 # audio file format detector
pydub==0.25.1  # audio file format converter
streamlit-audiorecorder==0.0.2 ## mic recorder

#TTS engine
pyttsx3==2.90
